pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "escaper"
version = "0.1.0"
description = "Experimental Social Collaborative Agent Platform for Escape-Room Reasoning"
authors = [
    {name = "Emile Al-Billeh"},
    {name = "Layanne A El Assaad"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]

[project.scripts]
escaper = "escaper.cli.run_experiment:main"

